`terraform` escape hatch properly checks/switches workspaces + stdout cleanup
It was confusing and problematic for the terraform escape hatch to avoid
doing a workspace check.  This is technically not backwards compatible
but I don't expect prior workflows to be affected negatively.

The environment-switching prompt needed to print to stderr, otherwise
mars could appear to hang if you were redirecting/capturing stdout.