INFR-18 upgrade kubectl
This moves the kubectl snap install back to the kubectl module and adds
conditional logic to prevent it being executed when the version is not
explicitly out of date.

This also uses a new pattern to prevent duplicate task execution across plays
from dependencies on the `kubectl` role.  This works by using a conditional
include that is overridden by setting a fact after the first run through the
full task list.