ansible vault Azure Key Vault support
This adds support for Azure Key Vault stored vault keys to the ansible wrapper
and refactors the wrapper to simplify it and improve extensibility.

This removes the `ansible-vault-{view,edit}` commands and reworks the
`ansible-vault-{encrypt,decrypt}` commands to use the ansible python libraries
directly and keep all keys in memory.  There is also a new
`ansible-vault-decrypt-key` command that directly decrypts a specific key from a
yaml file to avoid the need for copy paste.

The `ansible-playbook` command could not be easily implemented as an API call,
so it supports Key Vault via the ansible vault client script API using a script
that takes environment variables.  See:
https://docs.ansible.com/ansible/6/user_guide/vault.html#storing-passwords-in-third-party-tools-with-vault-password-client-scripts