move bastion python packages to virtualenv
This moves the python dependencies for the bastion into a virtualenv for
compatibility with new Python versions and better isolation from system Python
libraries.  To access the libraries, this resets the interpreter to point to the
virtualenv in the `kubectl` role which is already a dependency where the
libraries are needed.